Upgrade to Vite 5
I was trying to upgrade to Vite 5 in my project but the method of importing the svg url used in this example project (that mine uses as a base) causes a bug in production builds.

More details about the bug are available [on this issue](https://github.com/vitejs/vite/issues/15208).

This commit upgrades to Vite 5 and fixes the bug.